workde on DB

+++ models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
+from sqlalchemy import Boolean, Column, ForeignKey, Integer, String, Numeric
 from sqlalchemy.orm import relationship
 
 from .database import Base
@@ -8,5 +8,5 @@ class Item(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String, index=True)
-    price = Column(Integer, index=True)
+    price = Column(Numeric(5, 2), index=True)
     is_offer = Column(Boolean, index=True)
